branding-ai/requirements.txt
# ===============================
# FastAPI Core
# ===============================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# ===============================
# LangGraph / LangChain
# ===============================
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5

# ===============================
# OpenAI
# ===============================
openai>=1.10.0

# ===============================
# Database
# ===============================
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
# pymysql>=1.1.0

# ===============================
# Image Processing 🔥
# ===============================
Pillow>=10.0.0
cloudinary>=1.36.0

# ===============================
# Google 🔥 (이번 에러 해결용)
# ===============================
google-auth>=2.0.0
google-genai>=1.0.0

# ===============================
# Config / Utils
# ===============================
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0